fix ai

+++ ISU-Backend-CoreService/src/main/java/com/iseeyou/fortunetelling/dto/response/account/SimpleSeerCardResponse.java
@@ -15,7 +15,7 @@ public class SimpleSeerCardResponse {
     private String name;
     private String avatarUrl;
     private Double rating;
-    private Double totalRates;
+    private Integer totalRates;
     private String profileDescription;
     private List<String> specialities;
 }
